Se refactoriza el proyecto
